fixed unsafe use of CDC_Send(... buf, uint8_t count): count was only an uint8_t, passing arbitrary non 8-bit integers failed. use uint32_t instead - fixes 1899

+++ src/main/drivers/serial_usb_vcp.c
@@ -108,7 +108,7 @@ static void usbVcpWriteBuf(serialPort_t *instance, const void *data, int count)
 
 static bool usbVcpFlush(vcpPort_t *port)
 {
-    uint8_t count = port->txAt;
+    uint32_t count = port->txAt;
     port->txAt = 0;
 
     if (count == 0) {
